clean documentation
