+++ huginn-ebpf-common/src/lib.rs
@@ -14,10 +14,7 @@ pub mod syn_raw_v6;
 pub use syn_raw_v4::SynRawDataV4;
 pub use syn_raw_v6::SynRawDataV6;
 
-/// Const-evaluable string equality, for compile-time interface assertions.
-///
-/// Used by the kernel crate to tie its BPF entry-point fn names to the shared
-/// [`constants::XDP_SYN_PROGRAM`] / [`constants::TC_SYN_PROGRAM`] at build time.
+/// Compile-time string equality for BPF entry-point name assertions.
 #[inline]
 #[must_use]
 pub const fn str_eq(a: &str, b: &str) -> bool {
@@ -35,26 +32,13 @@ pub const fn str_eq(a: &str, b: &str) -> bool {
     true
 }
 
-/// Build the BPF map key from source IP and port (IPv4).
-///
-/// Both `src_ip` and `src_port` are in network byte order as read by the LE CPU.
-/// Userspace (`huginn-ebpf`) uses the same encoding in `make_bpf_key_v4`.
+/// BPF map key (IPv4). Network-byte-order IP and port as read on a LE CPU.
 #[inline(always)]
 pub fn make_key_v4(src_ip: u32, src_port: u16) -> u64 {
     ((src_ip as u64) << 16) | (src_port as u64)
 }
 
-/// Build the BPF map key from source IPv6 address and port (kernel side).
-///
-/// Returns an 18-byte array: the 16-byte address (raw packet bytes, already in NBO)
-/// followed by 2 bytes of port in native-endian order, which on a LE CPU recovers
-/// the original big-endian wire bytes stored in `tcp->source`.
-///
-/// `src_addr` is `ip6->saddr` as a byte array, no endianness conversion needed.
-/// `src_port` is `tcp->source` as a u16 NBO-as-NE value (raw LE read of BE bytes).
-///
-/// Userspace (`huginn-ebpf`) uses `make_bpf_key_v6` which reconstructs the same bytes
-/// from a host-order port via `to_be_bytes()`.
+/// BPF map key (IPv6). 16-byte addr + 2-byte port in wire byte order.
 #[inline(always)]
 pub fn make_key_v6(src_addr: [u8; 16], src_port: u16) -> [u8; 18] {
     let mut key = [0u8; 18];
