+++ huginn-ebpf-programs/src/signals/mod.rs
@@ -1,4 +1,3 @@
-//! Signal modules: each captures a specific observable (e.g. TCP SYN for OS fingerprinting).
-//! The XDP pipeline dispatches to each signal's handler after parsing the packet.
+//! Capture signals dispatched from the datapath pipelines.
 
 pub mod tcp_syn;
