+++ huginn-ebpf-programs/src/signals/tcp_syn/mod.rs
@@ -1,9 +1,4 @@
-//! TCP SYN signal: capture client SYN packets for OS fingerprinting (p0f-style).
-//!
-//! Contract with `huginn-ebpf`: map names (`tcp_syn_map_v4`, `tcp_syn_map_v6`,
-//! `syn_counter`, `syn_insert_failures`, `syn_captured`, `syn_malformed` and their
-//! `_v6` counterparts), layout of `SynRawDataV4`/`SynRawDataV6`, and key encoding via
-//! `make_key` / `make_key_v6` / `make_bpf_key` / `make_bpf_key_v6` must stay in sync.
+//! TCP SYN capture. Map names, value layout, and key encoding must match `huginn-ebpf`.
 
 mod handler;
 mod log_level;
