+++ huginn-ebpf-programs/src/tc/mod.rs
@@ -1,18 +1,4 @@
-//! TC (clsact) ingress capture pipeline for TCP SYN fingerprinting.
-//!
-//! Functional mirror of the XDP pipeline in `crate::xdp`, but over `TcContext`. TC ingress runs in
-//! the network stack and reads packet bytes via `bpf_skb_load_bytes` (`ctx.load`), which the kernel
-//! makes available even for **non-linear / GRO-merged** skbs. Returning `TC_ACT_OK` never drops the
-//! packet, so it works on VLAN/bond interfaces where generic XDP drops GRO-merged data packets. See
-//! `data/ebpf-vlan-tc-capture.md`.
-//!
-//! The capture logic (dst filter, SYN-no-ACK gate, quirks, map writes) is identical to XDP; only
-//! the packet-access layer changes. The map set, names, key encoding, and value layout are reused
-//! unchanged via `signals::tcp_syn`, keeping the on-disk contract byte-for-byte identical.
-//!
-//! Unlike `crate::xdp`, this pipeline is fully safe: `ctx.load` returns values by copy, and the
-//! loader-patched globals are read through the safe accessors in `signals::tcp_syn`. No `unsafe`
-//! appears here, so the module denies it outright.
+//! TC clsact ingress capture. GRO-safe alternative to XDP on VLAN/bond edges.
 #![deny(unsafe_code)]
 
 use aya_ebpf::programs::TcContext;
@@ -23,22 +9,14 @@ use crate::signals::tcp_syn;
 use huginn_ebpf_common::constants::*;
 use huginn_ebpf_common::headers::{EthHdr, Ip4Hdr, Ip6Hdr, TcpHdr, VlanHdr};
 
-/// TC ingress pipeline: parse L2/L3/L4 and dispatch TCP SYNs to the shared finishers.
-///
-/// At TC ingress the skb starts at the MAC header. On a VLAN sub-interface (e.g. `bond0.44`) the
-/// 802.1Q tag is already stripped, so `EthHdr.h_proto` is the inner ethertype; on trunk interfaces
-/// up to two VLAN tags are walked (parity with XDP). Any parse/bounds failure returns `Ok(())`,
-/// and the caller passes the packet (`TC_ACT_OK`) regardless.
 pub fn try_tc_syn(ctx: &TcContext) -> Result<(), ()> {
     let mut offset = 0usize;
 
-    // ── Ethernet ────────────────────────────────────────────────────────────────
     let eth: EthHdr = ctx.load(offset).map_err(|_| ())?;
     offset = offset.saturating_add(mem::size_of::<EthHdr>());
 
     let mut eth_type = eth.h_proto;
 
-    // Up to two VLAN tags (QinQ / 802.1ad). Won't trigger on a sub-interface (tag pre-stripped).
     if eth_type == ETH_P_8021Q || eth_type == ETH_P_8021AD {
         let vlan: VlanHdr = ctx.load(offset).map_err(|_| ())?;
         offset = offset.saturating_add(mem::size_of::<VlanHdr>());
@@ -60,7 +38,6 @@ pub fn try_tc_syn(ctx: &TcContext) -> Result<(), ()> {
     Ok(())
 }
 
-/// Parse IPv4 + TCP from the skb and dispatch a SYN to `finish_tcp_syn_v4`.
 fn handle_ipv4(ctx: &TcContext, offset: usize) -> Result<(), ()> {
     let ip: Ip4Hdr = ctx.load(offset).map_err(|_| ())?;
 
@@ -83,7 +60,6 @@ fn handle_ipv4(ctx: &TcContext, offset: usize) -> Result<(), ()> {
         return Ok(());
     }
 
-    // ── TCP ─────────────────────────────────────────────────────────────────────
     let tcp_offset = offset.saturating_add(ip_hdr_len);
     let tcp: TcpHdr = ctx.load(tcp_offset).map_err(|_| ())?;
 
@@ -121,10 +97,7 @@ fn handle_ipv4(ctx: &TcContext, offset: usize) -> Result<(), ()> {
     result.map_err(|_| ())
 }
 
-/// Parse IPv6 + TCP from the skb and dispatch a SYN to `finish_tcp_syn_v6`.
-///
-/// Like the XDP path, only packets whose fixed-header `nexthdr` is directly TCP are
-/// fingerprinted; extension headers before TCP are passed without fingerprinting.
+// Only fixed-header nexthdr == TCP is fingerprinted; extension headers before TCP are skipped.
 fn handle_ipv6(ctx: &TcContext, offset: usize) -> Result<(), ()> {
     let ip6: Ip6Hdr = ctx.load(offset).map_err(|_| ())?;
 
@@ -132,14 +105,12 @@ fn handle_ipv6(ctx: &TcContext, offset: usize) -> Result<(), ()> {
         return Ok(());
     }
 
-    // IPv6 destination address filter (all-zeros = accept any).
     let dst_ip_v6_val = tcp_syn::dst_ip_v6();
     let is_zero = dst_ip_v6_val.iter().all(|&b| b == 0);
     if !is_zero && ip6.daddr != dst_ip_v6_val {
         return Ok(());
     }
 
-    // ── TCP ─────────────────────────────────────────────────────────────────────
     let tcp_offset = offset.saturating_add(mem::size_of::<Ip6Hdr>());
     let tcp: TcpHdr = ctx.load(tcp_offset).map_err(|_| ())?;
 
@@ -177,14 +148,7 @@ fn handle_ipv6(ctx: &TcContext, offset: usize) -> Result<(), ()> {
     result.map_err(|_| ())
 }
 
-/// Read the TCP options block into a fixed 40-byte buffer, one byte at a time via `ctx.load::<u8>`.
-///
-/// We deliberately avoid `ctx.load_bytes(.., &mut buf)`: aya computes its length as
-/// `min(buf.len(), skb_len - offset)`, which the verifier sees as a *variable* that can be `0`, and
-/// `bpf_skb_load_bytes` rejects a zero-sized read (`R4 invalid zero-sized read`). Each `load::<u8>`
-/// is a **constant** 1-byte `bpf_skb_load_bytes`, and the loop is bounded by the fixed 40-byte array
-/// (`take(declared_optlen)`), so the program verifies. This mirrors the XDP path in
-/// `signals/tcp_syn/handler.rs`. The read stops at the first out-of-bounds byte (short skb).
+// load::<u8> per byte: bpf_skb_load_bytes rejects zero-length reads; the verifier accepts constant 1-byte loads.
 #[inline(always)]
 fn load_tcp_options(ctx: &TcContext, opts_offset: usize, tcp_hdr_len: usize) -> ([u8; 40], u8) {
     let declared_optlen = tcp_hdr_len
