+++ huginn-ebpf-programs/src/xdp/packet.rs
@@ -1,20 +1,11 @@
-//! Raw packet access for the XDP pipeline.
-//!
-//! XDP hands the program a pointer window (`ctx.data()`..`ctx.data_end()`); reading a header means
-//! bounds-checking an offset and casting. This is intrinsically `unsafe` and verifier-mandated, so
-//! it is confined to this module — the single `unsafe fn` below carries its own `#[allow]`.
+//! Bounds-checked packet pointer access for XDP.
 
 use aya_ebpf::programs::XdpContext;
 use core::mem;
 
-/// Returns a const pointer to `T` at `offset` bytes from the start of the
-/// packet, or `None` if the access would exceed `data_end`.
-///
 /// # Safety
 ///
-/// The caller may only dereference the returned pointer if `Some` was returned.
-/// Bounds are checked against `ctx.data_end()` before the cast; the verifier
-/// accepts this pattern.
+/// Caller may only dereference when `Some` is returned.
 #[allow(unsafe_code)]
 #[inline(always)]
 pub unsafe fn ptr_at<T>(ctx: &XdpContext, offset: usize) -> Option<*const T> {
