+++ huginn-ebpf/src/log_level.rs
@@ -1,17 +1,7 @@
-//! Verbosity control for the in-kernel `aya-log` datapath logging.
+//! In-kernel `aya-log` verbosity. Values match `log::LevelFilter`; patched into the `log_level` global at load time.
 
-/// Verbosity of the in-kernel `aya-log` datapath logging.
-///
-/// The numeric values match `log::LevelFilter` so the loader can patch the program's `log_level`
-/// global directly. A datapath macro at level *L* runs only when the configured level is `>= L`,
-/// so [`Off`](Self::Off) is genuinely zero-cost on the hot path and, e.g., [`Warn`](Self::Warn)
-/// avoids the per-SYN `debug!` write entirely.
-///
-/// The capture pipelines currently emit only `debug!` (on capture) and `warn!` (on map-insert
-/// failure); the other levels are accepted for completeness and simply never fire today.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
 pub enum EbpfLogLevel {
-    /// No logging (default). The `aya-log` code is compiled in but never executed.
     #[default]
     Off,
     Error,
@@ -22,7 +12,6 @@ pub enum EbpfLogLevel {
 }
 
 impl EbpfLogLevel {
-    /// Numeric encoding patched into the program's `log_level` global (matches `log::LevelFilter`).
     pub fn as_u8(self) -> u8 {
         match self {
             EbpfLogLevel::Off => 0,
@@ -34,8 +23,6 @@ impl EbpfLogLevel {
         }
     }
 
-    /// Parse a case-insensitive level name (`off`/`error`/`warn`/`info`/`debug`/`trace`).
-    /// Returns `None` for unknown values so callers can build their own error.
     pub fn parse(s: &str) -> Option<Self> {
         match s.trim().to_ascii_lowercase().as_str() {
             "off" | "none" | "disabled" => Some(EbpfLogLevel::Off),
@@ -48,7 +35,6 @@ impl EbpfLogLevel {
         }
     }
 
-    /// Lowercase canonical name, e.g. for defaulting `RUST_LOG` or logging the resolved config.
     pub fn as_str(self) -> &'static str {
         match self {
             EbpfLogLevel::Off => "off",
